Move fake build generation to the discord package
